refactor(ocale): extract createFakeClock to shared test helper
Two specs had drifting copies of a Date.now spy + time-advancing sleep
helper (rate-limit-queue.spec.ts and the new game-passes integration
suite). Moving it to tests/helpers/fake-clock.ts matches the existing
test-helper layout (fake-sleep.ts, fake-http-client.ts) and gives
future resource clients a single import instead of a third duplicate.
